requirements.txt
aiogram==3.13.1
pydantic==2.9.2
aiofiles==24.1.0
python-dotenv==1.0.1
Django==4.2.24
djangorestframework==3.16.1